Separate Serializer from file system logic
